+++ software/Server/Python/src/msxpi-server.py
@@ -20,7 +20,7 @@ from random import randint
 
 version = "0.9.0"
 build   = "20200810.00001"
-TRANSBLOCKSIZE = 1024
+BLKSIZE = 1024
 
 # Pin Definitons
 csPin   = 21
@@ -115,164 +115,104 @@ def SPI_MASTER_transfer_byte(byte_out):
     #print "transfer_byte:received",hex(byte_in),":",chr(byte_in)
     return byte_in
 
-def piexchangebyte(checktimeout,mypibyte):
-    time_start = time.time()
-    timeout = False
+def piexchangebyte(byte_out):
     rc = RC_SUCCESS
     
     GPIO.output(rdyPin, GPIO.HIGH)
-    #GPIO.wait_for_edge(csPin, GPIO.FALLING)
-    if (checktimeout):
-        while(GPIO.input(csPin)):
-            if (time.time() - time_start > BYTETRANSFTIMEOUT):
-                timeout = True
-    else:
-        while(GPIO.input(csPin)):
-            ()
+    while(GPIO.input(csPin)):
+        ()
 
-    mymsxbyte = SPI_MASTER_transfer_byte(mypibyte)
+    byte_in = SPI_MASTER_transfer_byte(byte_out)
     GPIO.output(rdyPin, GPIO.LOW)
 
-    if (timeout):
-        rc = RC_TIMEOUT
-
     #print "piexchangebyte: received:",hex(mymsxbyte)
-    return [rc,mymsxbyte]
+    return byte_in
 
-def sendstdmsg(rc, message):
-    piexchangebyte(NoTimeOutCheck,rc)
-    return senddatablock(TimeOutCheck,message,0,len(message),True)
-
-def recvdatablock(timeoutFlag):
+def recvdatablock():
     buffer = bytearray()
     bytecounter = 0
     crc = 0
     rc = RC_SUCCESS
     
-    mymsxbyte = piexchangebyte(timeoutFlag,SENDNEXT)
-    if (mymsxbyte[1] != SENDNEXT):
-        print "recvdatablock:Out of sync with MSX, waiting SENDNEXT, received rc,msxbyte:",hex(mymsxbyte[0]),hex(mymsxbyte[1])
+    msxbyte = piexchangebyte(SENDNEXT)
+    if (msxbyte != SENDNEXT):
+        print("recvdatablock:Out of sync with MSX")
         rc = RC_OUTOFSYNC
     else:
-        dsL = piexchangebyte(NoTimeOutCheck,SENDNEXT)
-        dsM = piexchangebyte(NoTimeOutCheck,SENDNEXT)
-        datasize = dsL[1] + 256 * dsM[1]
+        dsL = piexchangebyte(SENDNEXT)
+        dsM = piexchangebyte(SENDNEXT)
+        datasize = dsL + 256 * dsM
         
         #print "recvdatablock:Received blocksize =",datasize
-        while(datasize>bytecounter and rc == RC_SUCCESS):
-            mymsxbyte = piexchangebyte(TimeOutCheck,SENDNEXT)
-            if (mymsxbyte[0] == RC_SUCCESS):
-                #print "recvdatablock:Received byte:",chr(mymsxbyte[1])
-                buffer.append(mymsxbyte[1])
-                crc ^= mymsxbyte[1]
-                bytecounter += 1
-            else:
-                #print "recvdatablock:Error during transfer"
-                rc = RC_TIMEOUT
+        while(datasize>bytecounter):
+            msxbyte = piexchangebyte(SENDNEXT)
+            buffer.append(msxbyte)
+            crc ^= msxbyte
+            bytecounter += 1
 
-    if (rc == RC_SUCCESS):
-        #print "crc = ",crc
-        mymsxbyte = piexchangebyte(NoTimeOutCheck,crc)
-        if (mymsxbyte[1] != crc):
+        msxcrc = piexchangebyte(crc)
+        if (msxcrc != crc):
             rc = RC_CRCERROR
-        #else:
-            #print "recvdatablock:CRC verified"
 
     #print "recvdatablock:exiting with rc = ",hex(rc)
     return [rc,buffer]
 
-def secrecvdata(buffer,initbytepos):
-    rc = RC_SUCCESS
-    
-    #print "secrecvdata:starting"
-    
-    msxbyte = piexchangebyte(TimeOutCheck,SENDNEXT)
-    if (msxbyte[1]==SENDNEXT):
-        bytel = piexchangebyte(NoTimeOutCheck,SENDNEXT)[1]
-        bytem = piexchangebyte(NoTimeOutCheck,SENDNEXT)[1]
-        filesize = bytel + (bytem * 256)
+def sendstdmsg(rc, message):
+    piexchangebyte(rc)
+    senddatablock(message,len(message),0,1)
 
-        blocksize = filesize
-        if (filesize>512):
-            blocksize = 512
-        
-        index = 0
-        retries = 0
-        while(index<filesize):
-            retries = 0
-            rc = RC_UNDEFINED
-            while(retries < GLOBALRETRIES and rc <> RC_SUCCESS):
-                datablock = recvdatablock(NoTimeOutCheck)
-                rc = datablock[0]
-                retries += 1
-            
-            if(retries>GLOBALRETRIES):
-                break
+def senddatablock(buf,blocksize,blocknumber,attempts=1):
 
-            buffer[index+initbytepos:index+initbytepos+len(datablock[1])] = str(datablock[1])
-            index += 512
-            
-            if (filesize - index > 512):
-                blocksize = 512
-            else:
-                blocksize = filesize - index
-    
-        if(retries>=GLOBALRETRIES):
-            print "secrecvdata:Transfer interrupted due to CRC error"
-            rc = RC_CRCERROR
-        #else:
-            #print "secrecvdata:successful"
-                
-    else:
-        rc = RC_FAILNOSTD
-        print "secrecvdata:out of sync"
-        piexchangebyte(TimeOutCheck,rc)
-
-    #print "secrecvdata:Exiting with rc = ",hex(rc)
-    return rc
-
-def senddatablock(checktimeout,buffer,initpos,datasize,sendsize):
-    bytecounter = 0
     crc = 0
-    rc = RC_SUCCESS
+    rc = RC_FAILED
     
-    mymsxbyte = piexchangebyte(TimeOutCheck,SENDNEXT)
-    if (mymsxbyte[1] != SENDNEXT):
-        print "senddatablock:Out of sync with MSX, waiting SENDNEXT, received",hex(mymsxbyte[0]),hex(mymsxbyte[1])
-        rc = RC_OUTOFSYNC
+    bufsize = len(buf)
+    bufpos = blocksize*blocknumber
+
+    
+    if (blocksize <= bufsize - bufpos):
+        thisblocksize = blocksize
     else:
-        if (sendsize):
-            #print "senddatablock:Sending blocksize ",datasize
-            piexchangebyte(NoTimeOutCheck,datasize % 256)
-            piexchangebyte(NoTimeOutCheck,datasize / 256)
+        thisblocksize = bufsize < bufpos
+
+    if thisblocksize <= 0:
+        thisblocksize = 0
+
+    msxbyte = piexchangebyte(SENDNEXT)
+    if (msxbyte != SENDNEXT):
+        print "senddatablock:Out of sync with MSX, waiting SENDNEXT, received",hex(msxbyte),hex(msxbyte)
+        return RC_OUTOFSYNC
+    else:
+        piexchangebyte(thisblocksize % 256)
+        piexchangebyte(thisblocksize / 256)
+        if blocksize == 0:
+            return ENDTRANSFER
     
-        while(datasize>bytecounter and rc == RC_SUCCESS):
-            #mypibyte = ord(buffer[bytecounter])
-            mypibyte = ord(buffer[initpos+bytecounter])
-            #print "senddatablock:",mypibyte
-            rc = mymsxbyte[0]
-            mymsxbyte = piexchangebyte(TimeOutCheck,mypibyte)
-            if (rc == RC_SUCCESS):
-                #print "senddatablock:byte sent successfully"
-                crc ^= mypibyte
-                bytecounter += 1
-            else:
-                print "senddatablock:Error during transfer:",hex(rc)
-                rc = RC_TIMEOUT
+    piexchangebyte(attempts)
 
-    if (rc == RC_SUCCESS):
-        mymsxbyte = piexchangebyte(NoTimeOutCheck,crc)
+    while (attempts > 0 and rc != RC_SUCCESS):
+        bytecounter = 0
+        while(bytecounter < thisblocksize):
+            pibyte = ord(buf[bufpos+bytecounter])
+            piexchangebyte(pibyte)
+            crc ^= pibyte
+            bytecounter += 1
+
+        attempts -= 1
+
+        msxcrc = piexchangebyte(crc)
         #print "senddatablock:CRC local:remote = ",crc,":",mymsxbyte[1]
-        if (mymsxbyte[1] != crc):
+        if (msxcrc != crc):
             rc = RC_CRCERROR
-        #else:
-        #    print "senddatablock:CRC verified"
+        else:
+            rc = RC_SUCCESS
+    
+    return rc 
 
-    #print "senddatablock:Exiting with rc=",hex(rc)
-    return rc
 
 #   senddatablockC(TimeOutCheck,buffer,index+initpos,blocksize,True)
 def senddatablockC(flag1,buf,initpos,size,flag2):
+    print("senddatablock:starting")
     fh = open(RAMDISK+'/msxpi.tmp', 'wb')
     fh.write(buf[initpos:initpos+size])
     fh.flush()
@@ -285,55 +225,46 @@ def senddatablockC(flag1,buf,initpos,size,flag2):
     print "Exiting senddatablockC:call returned:",hex(rc)
     return rc
 
-def secsenddata(buffer, initpos, filesize):
-    rc = RC_SUCCESS
-    
-    #print "secsenddata:starting transfer for",filesize,"bytes"
-    
-    msxbyte = piexchangebyte(TimeOutCheck,SENDNEXT)
-    
-    if (msxbyte[1]==SENDNEXT):
-        piexchangebyte(NoTimeOutCheck,filesize % 256)
-        piexchangebyte(NoTimeOutCheck,filesize / 256)
-        
-        blocksize = filesize
-        if (filesize>512):
-            blocksize = 512
+def getpath(basepath, path):
 
-        index = 0
-        retries = 0
-        
-        while(index<filesize):
-            retries = 0
-            rc = RC_UNDEFINED
-            lastindex = index+blocksize
-                                
-            while(retries < GLOBALRETRIES and rc <> RC_SUCCESS):
-                #print "secsenddata:initpos=",index,"endpos=",lastindex,"retry=",retries
-                rc = senddatablock(TimeOutCheck,buffer,index+initpos,blocksize,True)
-                retries += 1
+    path=path.strip().rstrip(' \t\r\n\0')
+    if  path.startswith('/'):
+        urltype = 0 # this is an absolute local path
+        newpath = path
+    elif (path.startswith('http') or \
+          path.startswith('ftp') or \
+          path.startswith('nfs') or \
+          path.startswith('smb')):
+        urltype = 2 # this is an absolute network path
+        newpath = path
+    elif basepath.startswith('/'):
+        urltype = 1 # this is an relative local path
+        newpath = basepath + "/" + path
+    elif (basepath.startswith('http') or \
+          basepath.startswith('ftp') or \
+          basepath.startswith('nfs') or \
+          basepath.startswith('smb')):
+        urltype = 3 # this is an relative network path
+        newpath = basepath + "/" + path
 
-            if(retries>=GLOBALRETRIES):
-                break
-            else:
-                index += 512
-                if (filesize - index > 512):
-                    blocksize = 512
-                else:
-                    blocksize = filesize - index
-                                                                
-        if(retries>=GLOBALRETRIES):
-            print "secsenddata:Transfer interrupted due to CRC error"
-            rc = RC_CRCERROR
-        #else:
-            #print "secsenddata:successful"
-    else:
-        rc = RC_FAILNOSTD
-        print "secsenddata:out of sync"
-        piexchangebyte(TimeOutCheck,rc)
+    return [urltype, newpath]
 
-    #print "secsenddata:Exiting with rc = ",hex(rc)
-    return rc
+# create a subclass and override the handler methods
+class MyHTMLParser(HTMLParser):
+    def __init__(self):
+        self.reset()
+        self.NEWTAGS = []
+        self.NEWATTRS = []
+        self.HTMLDATA = []
+    def handle_starttag(self, tag, attrs):
+        self.NEWTAGS.append(tag)
+        self.NEWATTRS.append(attrs)
+    def handle_data(self, data):
+        self.HTMLDATA.append(data)
+    def clean(self):
+        self.NEWTAGS = []
+        self.NEWATTRS = []
+        self.HTMLDATA = []
 
 def msxdos_inihrd(filename, access=mmap.ACCESS_WRITE):
     #print "msxdos_inihrd:Starting"
@@ -346,21 +277,76 @@ def msxdos_inihrd(filename, access=mmap.ACCESS_WRITE):
 
     return rc
 
-def pdir(path):
-    #print("pdir:starting ",path)
+def dos83format(fname):
+    name = '        '
+    ext = '   '
 
-    send_byte(RC_WAIT)
-    GPIO.output(misoPin, GPIO.LOW)
+    finfo = fname.split('.')
+
+    name = str(finfo[0]).ljust(8)
+    if len(finfo) == 2:
+        ext = str(finfo[1]).ljust(3)
+    
+    return name+ext
+
+def ini_fcb(fname):
+
+    fpath = fname.split(':')
+    if len(fpath) == 1:
+        msxfile = str(fpath[0])
+        msxdrive = 0
+    else:
+        msxfile = str(fpath[1])
+        drvletter = str(fpath[0]).upper()
+        msxdrive = ord(drvletter) - 64
+
+    #convert filename to 8.3 format using all 11 positions required for the FCB
+    msxfcbfname = dos83format(msxfile)
+
+    #print("Drive, Filename:",msxdrive,msxfcbfname)
+
+    # send FCB structure to MSX
+    piexchangebyte(msxdrive)
+    for i in range(0,11):
+        piexchangebyte(ord(msxfcbfname[i]))
+        print(msxfcbfname[i]),
+
+def prun(cmd):
+    piexchangebyte(RC_WAIT)
     rc = RC_SUCCESS
+    
+    if (cmd.strip() == '' or len(cmd.strip()) == 0):
+        print "prun:syntax error"
+        sendstdmsg(RC_FAILED,"Syntax: prun <command> <::> command\nTo pipe a command to other, use :: instead of |")
+        rc = RC_FAILED
+    else:
+        cmd = cmd.replace('::','|')
+        try:
 
-    # basepath is global variable
-    # Need to extract command line parameters from the passed string
-    # Takes only 1st parameter, ignores everything else
-    # If no parameters was given, than assume current dirs
-    try:
+            p = Popen(cmd.decode(), shell=True, stdin=PIPE, stdout=PIPE, stderr=PIPE, close_fds=True)
+            buf = p.stdout.read()
+            if len(buf) == 0:
+                buf = "Pi:Command did not return any output.\n"
+
+            sendstdmsg(rc,buf)
+
+        except Exception as e:
+            rc = RC_FAILED
+            sendstdmsg(rc,"Pi:Error running the command.\n"+str(e)+'\n')
+
+    #print "prun:exiting rc:",hex(rc)
+    return rc
+
+def pdir(path):
+    msxbyte = piexchangebyte(RC_WAIT)
+    global psetvar
+    basepath = psetvar[0][1]
+    rc = RC_SUCCESS
+    #print "pdir:starting"
+
+    if (msxbyte == SENDNEXT):
         urlcheck = getpath(basepath, path)
         if (urlcheck[0] == 0 or urlcheck[0] == 1):
-            #print ("pdir:filesystem access:",urlcheck[1].decode())
             if (path.strip() == '*'):
                 prun('ls -l ' + urlcheck[1])
             elif ('*' in path):
@@ -372,31 +358,157 @@ def pdir(path):
             else:
                 prun('ls -l ' + urlcheck[1])
         else:
-            send_byte(RC_WAIT)
-            #print "pdir:network access:"+urlcheck[1].decode()
             parser = MyHTMLParser()
             try:
                 htmldata = urllib2.urlopen(urlcheck[1].decode()).read()
                 parser = MyHTMLParser()
                 parser.feed(htmldata)
                 buf = " ".join(parser.HTMLDATA)
-                send_byte(RC_SUCCESS)
-                if len(buf) == 0:
-                    buf = 'Empty directory'
-                rc = senddatablock(buf,0,len(buf))
+                piexchangebyte(RC_SUCCESS)
+                rc = senddatablock(buf,len(buf),0,1)
+
             except urllib2.HTTPError as e:
                 rc = RC_FAILED
                 print "pdir:http error "+ str(e)
-                send_byte(rc)
-                sendstdmsg(str(e))
-    except Exception as e:
-        send_byte(RC_FAILED)
-        print "pdir:http error "+ str(e)
-        sendstdmsg('pdir:'+str(e))
+                sendstdmsg(rc,str(e))
+    else:
+        rc = RC_FAILNOSTD
+        print "pdir:out of sync in RC_WAIT"
 
     #print "pdir:exiting rc:",hex(rc)
     return rc
+
+def pcd(path):    
+    msxbyte = piexchangebyte(RC_WAIT)
+    rc = RC_SUCCESS
+    global psetvar
+    basepath = psetvar[0][1]
+    newpath = basepath
     
+    #print "pcd:starting basepath:path=",basepath + ":" + path
+
+    if (msxbyte == SENDNEXT):
+        if (path == '' or path.strip() == "."):
+            sendstdmsg(rc,basepath+'\n')
+        elif (path.strip() == ".."):
+            newpath = basepath.rsplit('/', 1)[0]
+            if (newpath == ''):
+                newpath = '/'
+            psetvar[0][1] = newpath
+            sendstdmsg(rc,str(newpath+'\n'))
+        else:
+            #print "pcd:calling getpath"
+            urlcheck = getpath(basepath, path)
+            newpath = urlcheck[1]
+            print "pcd:getpath returned:",newpath
+            if (newpath[:4] == "http" or \
+                newpath[:3] == "ftp" or \
+                newpath[:3] == "nfs" or \
+                newpath[:3] == "smb"):
+                rc = RC_SUCCESS
+                psetvar[0][1] = newpath
+                sendstdmsg(rc,str(newpath+'\n'))
+            else:
+                newpath = str(newpath) #[:len(newpath)-1])
+                print "newpath=",type(newpath),len(newpath)
+                if (os.path.isdir(newpath)):
+                    psetvar[0][1] = newpath
+                    sendstdmsg(rc,newpath+'\n')
+                elif (os.path.isfile(str(newpath))):
+                    sendstdmsg(RC_FAILED,"Pi:Error - not a folder")
+                else:
+                    sendstdmsg(RC_FAILED,"Pi:Error - path not found")
+    else:
+        rc = RC_FAILNOSTD
+        print "pcd:out of sync in RC_WAIT"
+
+    return [rc, newpath]
+
+def pcopy(path=''):
+    piexchangebyte(RC_WAIT)
+    print("pcopy function")
+    buf = ''
+    rc = RC_SUCCESS
+
+    global psetvar
+    basepath = psetvar[0][1]
+    pcopyindex = 0
+    
+    if (path.startswith('http') or \
+        path.startswith('ftp') or \
+        path.startswith('nfs') or \
+        path.startswith('smb') or \
+        path.startswith('/')):
+        fileinfo = path
+    elif path == '':
+        fileinfo = basepath
+    else: 
+        fileinfo = basepath+'/'+path
+
+    fileinfo = fileinfo.split()
+
+    print("fileinfo, Types = ",fileinfo,type(basepath),type(path))
+
+    if len(fileinfo) == 1:
+        fname_rpi = str(fileinfo[0])
+        fname_msx_0 = fname_rpi.split('/')
+        fname_msx = str(fname_msx_0[len(fname_msx_0)-1])
+    elif len(fileinfo) == 2:
+        fname_rpi = str(fileinfo[0])
+        fname_msx = str(fileinfo[1])
+    else:
+        print("Pi:Command line parametrs invalid.")
+        sendstdmsg(RC_FAILED,"Pi:Command line parametrs invalid.")
+        return RC_FAILED
+
+    print("Pi:Reading:",basepath,"/",fname_rpi)
+
+    urlcheck = getpath(basepath, path)
+    # basepath 0 local filesystem
+    if (urlcheck[0] == 0 or urlcheck[0] == 1):
+        try:
+            with open(fname_rpi, mode='rb') as f:
+                buf = f.read()
+ 
+        except Exception as e:
+            print("pcopy:",str(e))
+            sendstdmsg(RC_FAILED,"Pi:"+str(e)+'\n') 
+            return RC_FAILED
+
+    else:
+        # basepath is network
+        print("pcopy:urlcheck[1]:",fname_rpi)
+        try:
+            parser = MyHTMLParser()
+            #creds = base64.encodestring('%s:%s' % (username, password)).replace('anonymous', 'anonymous@mail.com')
+            #parser.add_header("Authorization", "Basic %s" % creds)
+            htmldata = urllib2.urlopen(fname_rpi.decode()).read()
+            parser = MyHTMLParser()
+            parser.feed(htmldata)
+            buf = " ".join(parser.HTMLDATA)
+    
+        except Exception as e:
+            print "pcopy:http error "+ str(e)
+            sendstdmsg(RC_FAILED,"Pi:"+str(e))
+
+    if rc == RC_SUCCESS:
+        if filesize == 0:
+            sendstdmsg(RC_FILENOTFOUND,"Pi:No valid data found")
+        else:
+            msxbyte = piexchangebyte(RC_SUCCESS)
+            ini_fcb(fname_msx)         
+            blocknumber = 0   
+            while (rc == RC_SUCCESS):
+                #cmd = "sudo " + RAMDISK + "/uploaddata.msx " + RAMDISK + "/msxpi.tmp " + str(filesize) + " " + str(pcopyindex) + " " + str(GLOBALRETRIES)
+                #rc = subprocess.call(cmd, shell=True)
+                #init_spi_bitbang()
+                #GPIO.output(rdyPin, GPIO.LOW)
+                #print "pcopy:received from bufsend.c:",hex(rc) 
+                rc = senddatablock(buf,BLKSIZE,blocknumber,attempts=5)
+                if rc == RC_SUCCESS:
+                    blocknumber += 1
+    return rc
+
 """ ============================================================================
     msxpi-server.py
     main program starts here
@@ -450,14 +562,13 @@ print "Starting MSXPi Server Version ",version,"Build",build
 try:
     while True:
         print("st_recvcmd: waiting command")
-        rc = recvdatablock(NoTimeOutCheck)
-        print("Received command",rc)
+        rc = recvdatablock()
+        print"Received command",rc[1]
 
         if (rc[0] == RC_SUCCESS):
             try:
                 cmd = str(rc[1].split()[0]).lower()
                 parms = str(rc[1][len(cmd)+1:])
-                print("Received:",cmd,parms)
                 # Executes the command (first word in the string)
                 # And passes the whole string (including command name) to the function
                 # globals()['use_variable_as_function_name']()
